% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgam_plot.R
\name{dgam_plot}
\alias{dgam_plot}
\title{Plots probability density function of a gamma}
\usage{
dgam_plot(alpha, beta, rate = TRUE, by = 0.0025)
}
\arguments{
\item{alpha}{The shape parameter}

\item{beta}{The rate parameter if rate = TRUE, scale parameter otherwise}

\item{rate}{Sets beta to rate or shape}
}
\description{
Plots probability density function of a gamma
}
\examples{
dgam_plot(2,2)
}
