% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnb_plot.R
\name{dnb_plot}
\alias{dnb_plot}
\title{Plots probability mass function of a negative binomial
Number of failures until rth success}
\usage{
dnb_plot(r, p)
}
\arguments{
\item{r}{The final success}

\item{p}{Probability of success}
}
\description{
Plots probability mass function of a negative binomial
Number of failures until rth success
}
\examples{
dnb_plot(20,0.2)
}
