% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpois_plot.R
\name{dpois_plot}
\alias{dpois_plot}
\title{Plots probability mass function of a Poisson}
\usage{
dpois_plot(lambda)
}
\arguments{
\item{lambda}{The final success}
}
\description{
Plots probability mass function of a Poisson
}
\examples{
dpois_plot(10)
}
