% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcfunc.R
\name{gcfunc}
\alias{gcfunc}
\title{gcfunc}
\usage{
gcfunc(P)
}
\arguments{
\item{P}{is the p value from the result of GRAMMAR function}
}
\value{
completed value of the genomic control of p value
}
\description{
Perform genomic control of p value
}
\examples{
   P = as.matrix(read.table(Pvaluedata))
   gcp = gcfunc(P)
}
