% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRAMMAR.R
\name{GRAMMAR}
\alias{GRAMMAR}
\title{Generalized Renown Analysis of Molecular variance for Mixed model Analysis R package}
\usage{
GRAMMAR(X, Y, max_itr, num.parallel, outPath)
}
\arguments{
\item{X}{is the SNP matrix, individual x snp}

\item{Y}{is the phenotype matrix, individual x phenotype}

\item{max_itr}{is specifies the number of permutations.}

\item{num.parallel}{Number of parallel processes or a predefined socket cluster}

\item{outPath}{is the result file output path}
}
\value{
p-value and f-value
}
\description{
GRAMMAR function calculates p-value and f-value just one step
}
\examples{
   
   R CMD BATCH --args -Xpath= -Ypath= -max_itr= -num.parallel= -outPath=./ GRAMMAR.R GRAMMAR.log
   
}
