% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kinship.R
\name{Kinship}
\alias{Kinship}
\title{Kinship}
\usage{
Kinship(X)
}
\arguments{
\item{X}{is the SNP matrix, individual x snp}
}
\description{
Kinship function will estimate kinship coefficient
}
\examples{
   X = as.matrix(fread("SNP_rightdim.txt"))
   K = Kinship(t(X))
}
