% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRAMMAR.R
\name{run_grammar}
\alias{run_grammar}
\title{Generalized Renown Analysis of Molecular variance for Mixed model Analysis R package}
\usage{
run_grammar(K, Y, X, VC, max_itr, num.parallel, outPath)
}
\arguments{
\item{K}{is obtained from the Kinship function}

\item{Y}{is the phenotype matrix, individual x phenotype}

\item{X}{is the SNP matrix, individual x snp}

\item{VC}{is obtained from the varComp function}

\item{num.parallel}{Number of parallel processes or a predefined socket cluster}

\item{outPath}{is the file output path}

\item{mat_itr}{is specifies the number of permutations.}
}
\value{
p-value and f-value
}
\description{
run_grammar function calculates p-value and f-value.
}
\examples{
   X = as.matrix(fread("SNP_rightdim.txt"))
   Y = as.matrix(fread("Phenotype_rightdim.txt"))
   
   K = Kinship(t(X))
   VC = varComp(K, Y, X)
   
   result = run_grammar(K, Y, X, VC, max_itr = 4, num.parallel = 2, outPath = "./")
   
}
