% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varComp.R
\name{varComp}
\alias{varComp}
\title{varComp}
\usage{
varComp(K, Y, X)
}
\arguments{
\item{K}{is obtained from the Kinship function}

\item{Y}{is the phenotype matrix, individual x phenotype}

\item{X}{is the SNP matrix, individual x snp}
}
\value{
Variance components Vg: genetic factor(VC[1]) Ve: environment factor(VC[2])

UY is the third output value of the varComp function.

UX is the fourth output value of the varComp function.
}
\description{
varComp function will estimate variance components
}
\examples{
   X = as.matrix(fread("X_rightdim.txt"))
   K = Kinship(t(X))
   Y = as.matrix(fread("Y_rightdim.txt"))
   VC = varComp(K, Y, X)
}
