% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRAMMAR.R
\name{run_grammar}
\alias{run_grammar}
\title{Generalized Renown Analysis of Molecular variance for Mixed model Analysis R package}
\usage{
run_grammar(UY, UX, max_itr, num.parallel)
}
\arguments{
\item{UY}{is obtained from the varComp function}

\item{UX}{is obtained from the varComp function}

\item{num.parallel}{Number of parallel processes or a predefined socket cluster}

\item{K}{is obtained from the Kinship function}

\item{mat_itr}{is specifies the number of permutations.}
}
\value{
p-value and f-value
}
\description{
run_gamma function calculates p-value and f-value.
}
\examples{
   X = as.matrix(fread("SNP_rightdim.txt"))
   K = Kinship(t(X))
   
   Y = as.matrix(fread("Phenotype_rightdim.txt"))
   VC = varComp(K, Y, X)
   
   ps = run_grammar(VC[3], VC[4], max_itr = 4, num.parallel = 2)
   
   # ps[1] = p-value
   # ps[2] = f-value
}
