% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varComp.R
\name{varComp}
\alias{varComp}
\title{varComp}
\usage{
varComp(Y, K, num.parallel)
}
\arguments{
\item{Y}{is the phenotype matrix, individual x phenotypes}

\item{K}{is obtained from the Kinship function}

\item{num.parallel}{Number of parallel processes or a predefined socket cluster}
}
\value{
Variance components Vg: genetic factor Ve: environment factor
}
\description{
varComp function will estimate variance components
}
\examples{
   X <- as.matrix(data.table::fread(Genotypesdata))
   Y <- as.matrix(data.table::fread(Phenotypesdata))
   K <- Kinship(X)
   
   VC <- varComp(Y, K, num.parallel)
}
