% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varComp.R
\name{varComp}
\alias{varComp}
\title{varComp}
\usage{
varComp(K, Y, X)
}
\arguments{
\item{K}{is obtained from the Kinship function}

\item{Y}{is the phenotype matrix, individual x phenotypes}

\item{X}{is the SNP matrix, individual x SNPs}
}
\value{
Variance components Vg: genetic factor Ve: environment factor
}
\description{
varComp function will estimate variance components
}
\examples{
   X = as.matrix(read.table("./X_rightdim.txt"))
   Y = as.matrix(read.table("./Y_rightdim.txt"))
   K = Kinship(t(X))
   
   VC = varComp(K, Y, X)
}
